requirements.txt
asgiref==3.2.10
azure-ai-textanalytics==5.1.0b2
azure-common==1.1.25
azure-core==1.8.2
certifi==2020.6.20
chardet==3.0.4
datefinder==0.7.1
Django==3.1.2
djangorestframework==3.12.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
word2number==1.1
